# field-service/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.FieldResponse:
    properties:
      code:
        type: string
      created_at:
        type: string
      images:
        items:
          type: string
        type: array
      name:
        type: string
      price_per_hour:
        type: integer
      updated_at:
        type: string
      uuid:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      message: {}
      status:
        type: string
      token:
        type: string
    type: object
  util.PaginationResult:
    properties:
      data: {}
      limit:
        type: integer
      next_page:
        type: integer
      page:
        type: integer
      previous_page:
        type: integer
      total_data:
        type: integer
      total_page:
        type: integer
    type: object
host: localhost:8002
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A microservice for managing booking system fields, operational times,
    and schedules.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Field Service API
  version: "1.0"
paths:
  /field:
    get:
      consumes:
      - application/json
      description: Get a list of all fields
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.FieldResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get all fields without pagination
      tags:
      - fields
    post:
      consumes:
      - multipart/form-data
      description: Create a new field with images
      parameters:
      - description: Field Name
        in: formData
        name: name
        required: true
        type: string
      - description: Field Code
        in: formData
        name: code
        required: true
        type: string
      - description: Price Per Hour
        in: formData
        name: price_per_hour
        required: true
        type: integer
      - description: Field Images
        in: formData
        name: images
        required: true
        type: file
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FieldResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Create a new field
      tags:
      - fields
  /field/{uuid}:
    delete:
      consumes:
      - application/json
      description: Delete a field by UUID
      parameters:
      - description: Field UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Delete a field
      tags:
      - fields
    get:
      consumes:
      - application/json
      description: Get detailed information about a field
      parameters:
      - description: Field UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FieldResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get field by UUID
      tags:
      - fields
    put:
      consumes:
      - multipart/form-data
      description: Update field details and images
      parameters:
      - description: Field UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Field Name
        in: formData
        name: name
        required: true
        type: string
      - description: Field Code
        in: formData
        name: code
        required: true
        type: string
      - description: Price Per Hour
        in: formData
        name: price_per_hour
        required: true
        type: integer
      - description: Field Images
        in: formData
        name: images
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/dto.FieldResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Update an existing field
      tags:
      - fields
  /field/pagination:
    get:
      consumes:
      - application/json
      description: Get a paginated list of fields
      parameters:
      - description: Page number
        in: query
        minimum: 1
        name: page
        required: true
        type: integer
      - description: Number of items per page
        in: query
        minimum: 1
        name: limit
        required: true
        type: integer
      - description: Column to sort by
        enum:
        - id
        - date
        - time
        - status
        - created_at
        - updated_at
        in: query
        name: sort_column
        required: true
        type: string
      - description: Sort order
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/util.PaginationResult'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get fields with pagination
      tags:
      - fields
...
schemes:
- http
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"